# Swagger_v4.YAML
openapi: 3.0.0
info:
  title: API Sistema de Subastas
  version: 1.2.0
  description: Documentacion de la API REST para la aplicacion movil del Sistema de Subastas.
  contact:
    name: Equipo de Desarrollo
...
paths:
...
  /usuarios/me:
    get:
      summary: Obtener el perfil del usuario autenticado
      tags: [Perfil]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: No autorizado.

    patch:
      summary: Actualizar datos del perfil
      tags: [Perfil]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UsuarioUpdate"
      responses:
        "200":
          description: Perfil actualizado exitosamente.
        "401":
          description: No autorizado.
  /usuarios/me/medios-pago:
    get:
      summary: Listar medios de pago
      tags: [Perfil]
      description: Muestra cuentas bancarias, tarjetas o cheques registrados por el usuario.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de medios de pago.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedioPago"
        "401":
          description: No autorizado.
    post:
      summary: Agregar medio de pago
      tags: [Perfil]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedioPagoInput"
      responses:
        "201":
          description: Medio de pago agregado. Queda pendiente de verificacion.
        "401":
          description: No autorizado.
  /usuarios/me/medios-pago/{id}:
    patch:
      summary: Actualizar medio de pago
      tags: [Perfil]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedioPagoUpdate"
      responses:
        "200":
          description: Medio de pago actualizado.
        "401":
          description: No autorizado.
        "404":
          description: Medio de pago no encontrado.
    delete:
      summary: Eliminar medio de pago
      tags: [Perfil]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Medio de pago eliminado.
        "401":
          description: No autorizado.
        "404":
          description: Medio de pago no encontrado.
  /usuarios/me/metricas:
    get:
      summary: Listar metricas del perfil
      tags: [Perfil]
      description: Metricas de participacion, ofertas, importes y categorias.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista las metricas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioMetricas"
        "401":
          description: No autorizado.
  /usuarios/me/multas:
    get:
      summary: Listar multas activas
      tags: [Perfil]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Multas del usuario.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Multa"
        "401":
          description: No autorizado.
  /usuarios/me/multas/pagar:
    post:
      summary: Pagar multa
      tags: [Perfil]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultaPagoRequest"
      responses:
        "200":
          description: Multa pagada.
        "400":
          description: Medio de pago invalido.
        "401":
          description: No autorizado.
        "404":
          description: Multa no encontrada.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Categoria:
      type: string
      enum: [comun, especial, plata, oro, platino]
    Moneda:
      type: string
      enum: [ARS, USD]
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
        documento:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        direccion:
          type: string
        foto:
          type: string
          format: uri
          example: "https://storage.example.com/users/35123456/photo.jpg"
        numeroPais:
          type: integer
        admitido:
          type: string
          enum: [si, no]
        estadoRegistro:
          type: string
          enum: [pendiente, aprobado, rechazado]
        categoria:
          $ref: "#/components/schemas/Categoria"
        multaActiva:
          type: boolean
          description: Si el usuario tiene una multa del 10% sin pagar.
        bloqueado:
          type: boolean
          description: Si el usuario quedo bloqueado por incumplimiento.
    UsuarioUpdate:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        direccion:
          type: string
        telefono:
          type: string
        foto:
          type: string
          format: binary
...
    UsuarioMetricas:
      type: object
      properties:
        totalSubastasParticipadas:
          type: integer
          example: 15
        totalSubastasGanadas:
          type: integer
          example: 4
        porcentajeExito:
          type: number
          format: float
          example: 26.6
        totalPujasRealizadas:
          type: integer
          example: 120
        montoTotalOfertado:
          type: number
          format: float
          example: 350000.50
        montoTotalPagado:
          type: number
          format: float
          example: 120000.00
        categoriasParticipadas:
          type: array
          items:
            $ref: "#/components/schemas/Categoria"
        ultimaParticipacion:
          type: string
          format: date-time
    MedioPago:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [tarjeta_credito, cuenta_bancaria, cheque_certificado]
        ultimos_digitos:
          type: string
          example: "4321"
        estadoVerificacion:
          type: string
          enum: [pendiente, validado, rechazado]
        moneda:
          $ref: "#/components/schemas/Moneda"
        limiteReservado:
          type: number
          format: float
          example: 50000.00
        paisBanco:
          type: string
          example: "AR"
        esCuentaReceptora:
          type: boolean
          description: Indica si esta cuenta se usa para recibir pagos.
    MedioPagoInput:
      type: object
      required: [tipo, datos_encriptados, moneda]
      properties:
        tipo:
          type: string
          enum: [tarjeta_credito, cuenta_bancaria, cheque_certificado]
        datos_encriptados:
          type: string
          description: Token generado por la pasarela de pagos.
        moneda:
          $ref: "#/components/schemas/Moneda"
        limiteReservado:
          type: number
          format: float
        paisBanco:
          type: string
        esCuentaReceptora:
          type: boolean
    MedioPagoUpdate:
      type: object
      properties:
        limiteReservado:
          type: number
          format: float
        esCuentaReceptora:
          type: boolean
...
    Multa:
      type: object
      properties:
        id:
          type: integer
        importe:
          type: number
          format: float
        estado:
          type: string
          enum: [pendiente, pagada]
        fechaLimite:
          type: string
          format: date-time
        motivo:
          type: string
    MultaPagoRequest:
      type: object
      required: [multaId, medioPagoId]
      properties:
        multaId:
          type: integer
        medioPagoId:
          type: integer
...